change nlu to structured output
