+++ arklex/orchestrator/NLU/services/model_service.py
@@ -12,7 +12,6 @@ from typing import Any
 from langchain_core.language_models import BaseChatModel
 from langchain_core.messages import HumanMessage, SystemMessage
 
-from arklex.orchestrator.NLU.services.api_service import APIClientService
 from arklex.orchestrator.NLU.utils.formatters import (
     format_verification_input as format_verification_input_formatter,
 )
@@ -55,15 +54,7 @@ class ModelService:
         self.model_config = model_config
         self._validate_config()
         try:
-            self.api_service = APIClientService(base_url=self.model_config["endpoint"])
             self.model = self._initialize_model()
-            log_context.info(
-                "ModelService initialized successfully",
-                extra={
-                    "model_name": model_config.get("model_name"),
-                    "operation": "initialization",
-                },
-            )
         except Exception as e:
             log_context.error(
                 LOG_MESSAGES["ERROR"]["INITIALIZATION_ERROR"].format(
@@ -179,107 +170,6 @@ class ModelService:
                     "operation": "model_initialization",
                 },
             ) from e
-        
-    def _format_intent_definition(
-        self, intent_name: str, definition: str, count: int
-    ) -> str:
-        """Format a single intent definition.
-
-        Args:
-            intent_name: Name of the intent
-            definition: Intent definition text
-            count: Intent number in sequence
-
-        Returns:
-            Formatted intent definition string
-        """
-        return f"{count}) {intent_name}: {definition}\n"
-
-    def _format_intent_exemplars(
-        self, intent_name: str, sample_utterances: list[str], count: int
-    ) -> str:
-        """Format sample utterances for an intent.
-
-        Args:
-            intent_name: Name of the intent
-            sample_utterances: List of example utterances
-            count: Intent number in sequence
-
-        Returns:
-            Formatted exemplars string
-        """
-        if not sample_utterances:
-            return ""
-        exemplars = "\n".join(sample_utterances)
-        return f"{count}) {intent_name}: \n{exemplars}\n"
-
-    def _process_intent(
-        self,
-        intent_k: str,
-        intent_v: list[dict[str, Any]],
-        count: int,
-        idx2intents_mapping: dict[str, str],
-    ) -> tuple[str, str, str, int]:
-        """Process a single intent and its variations.
-
-        Args:
-            intent_k: Intent key/name
-            intent_v: List of intent definitions
-            count: Current count for numbering
-            idx2intents_mapping: Mapping of indices to intent names
-
-        Returns:
-            Tuple containing:
-                - definition_str: Formatted definitions
-                - exemplars_str: Formatted exemplars
-                - intents_choice: Formatted choices
-                - new_count: Updated count
-        """
-        definition_str = ""
-        exemplars_str = ""
-        intents_choice = ""
-
-        if len(intent_v) == 1:
-            intent_name = intent_k
-            idx2intents_mapping[str(count)] = intent_name
-            definition = intent_v[0].get("attribute", {}).get("definition", "")
-            sample_utterances = (
-                intent_v[0].get("attribute", {}).get("sample_utterances", [])
-            )
-
-            if definition:
-                definition_str += self._format_intent_definition(
-                    intent_name, definition, count
-                )
-            if sample_utterances:
-                exemplars_str += self._format_intent_exemplars(
-                    intent_name, sample_utterances, count
-                )
-            intents_choice += f"{count}) {intent_name}\n"
-
-            count += 1
-        else:
-            for idx, intent in enumerate(intent_v):
-                intent_name = f"{intent_k}__<{idx}>"
-                idx2intents_mapping[str(count)] = intent_name
-                definition = intent.get("attribute", {}).get("definition", "")
-                sample_utterances = intent.get("attribute", {}).get(
-                    "sample_utterances", []
-                )
-
-                if definition:
-                    definition_str += self._format_intent_definition(
-                        intent_name, definition, count
-                    )
-                if sample_utterances:
-                    exemplars_str += self._format_intent_exemplars(
-                        intent_name, sample_utterances, count
-                    )
-                intents_choice += f"{count}) {intent_name}\n"
-
-                count += 1
-
-        return definition_str, exemplars_str, intents_choice, count
 
     def get_response(
         self,
@@ -330,114 +220,25 @@ class ModelService:
     def get_response_with_structured_output(
         self,
         prompt: str,
-        model_config: dict[str, Any] | None = None,
         schema: dict[str, Any] | None = None,
         system_prompt: str | None = None,
     ) -> str:
         """Get response from the model with structured output."""
         # Check if the model is an OpenAI model by checking the model_config
         is_openai_model = (
-            self.model_config.get("llm_provider", "").lower() == "openai" or
-            "openai" in str(self.model).lower()
+            self.model_config.get("llm_provider", "").lower() == "openai"
+            or "openai" in str(self.model).lower()
         )
-        
+
         if is_openai_model:
             messages = []
             if system_prompt:
                 messages.append(SystemMessage(content=system_prompt))
             messages.append(HumanMessage(content=prompt))
             llm = self.model.with_structured_output(schema)
-            
             return llm.invoke(messages)
         else:
             return self.get_response(prompt, system_prompt)
-    
-
-    def get_json_response(
-        self,
-        prompt: str,
-        model_config: dict[str, Any] | None = None,
-        system_prompt: str | None = None,
-    ) -> dict[str, Any]:
-        """Get JSON response from the model.
-
-        Sends a prompt to the model and returns its response as a parsed
-        JSON object. Handles message formatting and JSON validation.
-
-        Args:
-            prompt: User prompt to send to the model
-            model_config: Optional model configuration parameters. If not provided,
-                         uses the instance's model_config.
-            system_prompt: Optional system prompt for model context
-
-        Returns:
-            Parsed JSON response
-
-        Raises:
-            ValueError: If JSON parsing fails or response is invalid
-        """
-        try:
-            response = self.get_response(prompt, system_prompt)
-            return json.loads(response)
-        except json.JSONDecodeError as e:
-            log_context.error(f"Error parsing JSON response: {str(e)}")
-            raise ValueError(f"Failed to parse JSON response: {str(e)}") from e
-        except Exception as e:
-            log_context.error(f"Error getting JSON response: {str(e)}")
-            raise ValueError(f"Failed to get JSON response: {str(e)}") from e
-
-    def format_intent_input(
-        self, intents: dict[str, list[dict[str, Any]]], chat_history_str: str
-    ) -> tuple[str, dict[str, str]]:
-        """Format input for intent detection.
-
-        Creates a formatted prompt for intent detection based on the
-        provided intents and chat history. Also generates a mapping
-        from indices to intent names.
-
-        Args:
-            intents: Dictionary of intents containing:
-                - intent_name: List of intent definitions
-                - attribute: Intent attributes (definition, sample_utterances)
-            chat_history_str: Formatted chat history
-
-        Returns:
-            Tuple containing:
-                - formatted_prompt: Formatted prompt for intent detection
-                - idx2intents_mapping: Mapping from indices to intent names
-        """
-        definition_str = ""
-        exemplars_str = ""
-        intents_choice = ""
-        idx2intents_mapping: dict[str, str] = {}
-        count = 1
-
-        for intent_k, intent_v in intents.items():
-            def_str, ex_str, choice_str, new_count = self._process_intent(
-                intent_k, intent_v, count, idx2intents_mapping
-            )
-            definition_str += def_str
-            exemplars_str += ex_str
-            intents_choice += choice_str
-            count = new_count
-
-        prompt = f"""Given the following intents and their definitions, determine the most appropriate intent for the user's last input.
-
-Intent Definitions:
-{definition_str}
-
-Sample Utterances:
-{exemplars_str}
-
-Available Intents:
-{intents_choice}
-
-Chat History:
-{chat_history_str}
-
-Please choose the most appropriate intent by providing the corresponding intent number and intent name in the format of 'intent_number) intent_name'."""
-
-        return prompt, idx2intents_mapping
 
     def format_slot_input(
         self, slots: list[dict[str, Any]], context: str, type: str = "chat"
@@ -683,26 +484,3 @@ class DummyModelService(ModelService):
             Tuple[bool, str]: Verification result and explanation
         """
         return super().process_verification_response(response)
-
-    def get_json_response(
-        self,
-        prompt: str,
-        model_config: dict[str, Any] | None = None,
-        system_prompt: str | None = None,
-    ) -> dict[str, Any]:
-        """Get a mock JSON response for testing.
-
-        Args:
-            prompt: Input prompt
-            model_config: Optional model configuration
-            system_prompt: Optional system prompt
-
-        Returns:
-            dict[str, Any]: Mock JSON response for testing
-        """
-        # Handle None or empty prompts
-        if prompt is None:
-            prompt = ""
-
-        # Return appropriate mock JSON responses based on the input
-        return {"result": "mock_response", "status": "success"}
